removed range estimations from conversions and them to a new module range estimations which has a new enum for Slant range and GSMR
